Only show update snackbar on Home
