+++ packages/toolpad-app/src/constants.ts
@@ -7,3 +7,4 @@ export const RUNTIME_CONFIG_WINDOW_PROPERTY = '__TOOLPAD_RUNTIME_CONFIG__';
 export const TOOLPAD_TARGET_CE = 'CE';
 export const TOOLPAD_TARGET_CLOUD = 'CLOUD';
 export const TOOLPAD_TARGET_PRO = 'PRO';
+export const DOCUMENTATION_URL = 'https://mui.com/toolpad/getting-started/setup/';
