+++ packages/toolpad-app/src/toolpad/Home.tsx
@@ -17,6 +17,7 @@ import {
   MenuItem,
   Skeleton,
   Stack,
+  Snackbar,
   Table,
   TableBody,
   TableCell,
@@ -30,6 +31,8 @@ import {
 } from '@mui/material';
 import { LoadingButton } from '@mui/lab';
 import IconButton from '@mui/material/IconButton';
+import CloseIcon from '@mui/icons-material/Close';
+import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
 import DriveFileRenameOutlineIcon from '@mui/icons-material/DriveFileRenameOutline';
 import ViewListIcon from '@mui/icons-material/ViewList';
@@ -49,6 +52,9 @@ import ErrorAlert from './AppEditor/PageEditor/ErrorAlert';
 import { ConfirmDialog } from '../components/SystemDialogs';
 import config from '../config';
 import { errorFrom } from '../utils/errors';
+import { DOCUMENTATION_URL, TOOLPAD_TARGET_CLOUD } from '../constants';
+
+const CURRENT_RELEASE_VERSION = `v1${process.env.TOOLPAD_VERSION}`;
 
 export interface CreateAppDialogProps {
   open: boolean;
@@ -469,6 +475,72 @@ function AppsListView({ loading, apps, activeDeploymentsByApp, setDeletedApp }: 
   );
 }
 
+function UpdateBanner() {
+  const { data: latestRelease } = client.useQuery('getLatestToolpadRelease', [], {
+    staleTime: 1000 * 60 * 10,
+    enabled: process.env.TOOLPAD_TARGET !== TOOLPAD_TARGET_CLOUD,
+  });
+
+  const [dismissedVersion, setDismissedVersion] = useLocalStorageState<string | null>(
+    'update-banner-dismissed-version',
+    null,
+  );
+
+  const handleDismissClick = React.useCallback(
+    (event: React.SyntheticEvent | Event, reason?: string) => {
+      if (reason === 'clickaway' || reason === 'escapeKeyDown') {
+        return;
+      }
+      setDismissedVersion(CURRENT_RELEASE_VERSION);
+    },
+    [setDismissedVersion],
+  );
+
+  const hideBanner =
+    (latestRelease && latestRelease.tag === CURRENT_RELEASE_VERSION) ||
+    dismissedVersion === CURRENT_RELEASE_VERSION;
+
+  return latestRelease ? (
+    <Snackbar
+      open={!hideBanner}
+      onClose={handleDismissClick}
+      anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
+      message={
+        <React.Fragment>
+          A new version <strong>{latestRelease.tag}</strong> of Toolpad is available.
+        </React.Fragment>
+      }
+      action={
+        <Stack direction="row" sx={{ gap: 2 }}>
+          <Button
+            aria-label="update"
+            color="inherit"
+            endIcon={<OpenInNewIcon fontSize="inherit" />}
+            component="a"
+            target="_blank"
+            href={DOCUMENTATION_URL}
+          >
+            Update
+          </Button>
+          <Button
+            aria-label="view changelog"
+            color="inherit"
+            endIcon={<OpenInNewIcon fontSize="inherit" />}
+            component="a"
+            target="_blank"
+            href={latestRelease.url}
+          >
+            View changelog
+          </Button>
+          <IconButton aria-label="close" color="inherit" size="small" onClick={handleDismissClick}>
+            <CloseIcon fontSize="inherit" />
+          </IconButton>
+        </Stack>
+      }
+    />
+  ) : null;
+}
+
 export default function Home() {
   const { data: apps = [], isLoading, error } = client.useQuery('getApps', []);
   const { data: activeDeployments } = client.useQuery('getActiveDeployments', []);
@@ -535,6 +607,7 @@ export default function Home() {
             setDeletedApp={setDeletedApp}
           />
         )}
+        <UpdateBanner />
       </Container>
     </ToolpadShell>
   );
